+++ src/ExportMap.js
@@ -645,7 +645,7 @@ ExportMap.parse = function (path, content, context) {
     // This doesn't declare anything, but changes what's being exported.
     if (includes(exports, n.type)) {
       const exportedName = n.type === 'TSNamespaceExportDeclaration'
-        ? n.id.name
+        ? (n.id || n.name).name
         : (n.expression && n.expression.name || (n.expression.id && n.expression.id.name) || null);
       const declTypes = [
         'VariableDeclaration',